convert into navbar compontent
